push script gradle & modif pull request

+++ serveur-application/vagrant-rsync/install_python_git.sh
@@ -22,10 +22,18 @@ apt install python3 python3-dev python3-pip git -q -y
 apt remove -y python 
 
 # positionnement de python3 et pip3 dans le profil
-
-ln -s /usr/bin/python3 /usr/bin/python
+ps_verif_dossier_pyhton() {
+    
+    if [ -f "/usr/bin/python" ] ; then
+        echo "file exist"
+    else
+        ln -s /usr/bin/python3 /usr/bin/python
+    fi
+}
+ps_verif_dossier_pyhton
 echo "alias pip=pip3" > ~/.bashrc
-
 #install utilitaire de test dont flask et fastapi pour expostion de l'api
 pip install flask pytest 
 pip install fastapi uvicorn
+
+echo "Success"
